+++ RyuSocks/Commands/Command.cs
@@ -34,6 +34,22 @@ namespace RyuSocks.Commands
             throw new NotSupportedException("This command does not require a second connection, so this method must not be called.");
         }
 
+        public virtual void GetSocketOption(
+            SocketOptionLevel optionLevel,
+            SocketOptionName optionName,
+            byte[] optionValue)
+        {
+            throw new NotSupportedException("This command does not require a second connection, so this method must not be called.");
+        }
+
+        public virtual void SetSocketOption(
+            SocketOptionLevel optionLevel,
+            SocketOptionName optionName,
+            byte[] optionValue)
+        {
+            throw new NotSupportedException("This command does not require a second connection, so this method must not be called.");
+        }
+
         public virtual int Send(ReadOnlySpan<byte> buffer, SocketFlags socketFlags, out SocketError errorCode)
         {
             throw new NotSupportedException("This command does not require a second connection, so this method must not be called.");